fix: user router fix
