+++ src/server/models/content.model.ts
