+++ src/server/types/user.type.ts
@@ -1,0 +1,7 @@
+export interface User {
+  first_name: string;
+  last_name: string;
+  username: string;
+  email: string;
+  password: string;
+}
